lcw
1、消除友盟xib警告
2、三方相册JKPhotoBrowser预览模糊
3、三方相册图片显示顺序，最新放最前边
4、修复LPhotoBrowse显示特长图片不能屏幕宽度显示问题
5、上传报告时，先截取尺寸375，再大小压缩

+++ TiJian/TiJian/HelpTools/MiddleTools.m
@@ -421,7 +421,7 @@
     [MBProgressHUD showHUDAddedTo:controller.view animated:YES];
     [[SDWebImageManager sharedManager]downloadImageWithURL:url options:SDWebImageRetryFailed progress:^(NSInteger receivedSize, NSInteger expectedSize) {
         
-        DDLOG(@"%ld %ld",receivedSize,expectedSize);
+        DDLOG(@"%ld %ld",(long)receivedSize,expectedSize);
         
     } completed:^(UIImage *image, NSError *error, SDImageCacheType cacheType, BOOL finished, NSURL *imageURL) {
